Add (functional) coverage and static analysis badges
Coveralls wasn't working at all, so I went with Codecov for a gcov-based coverage badge. It just seems easier to work with.

Also added Codacy for an approximate "code quality" badge.

I'm using gcov in "make check" to make sure that coverage stays above 90% for a check to succeed.

Also, the last stage of "make check" uses clang-tidy for static analysis such that any clang-tidy warnings cause the check to fail with an error.