This seems to be a poorly-phrased problem statement, or perhaps I'm missing some critical piece of information.
The test case below fails because a minimum height of 15 is expected, but my solution ends with a minimum height of 12.

books = {{7,3},{8,7},{2,7},{2,5}}
shelf_width = 10

My solution is definitely brute force O(N^2) time (at least).

This is likely a DP problem that could be solved in something like O(N) time and O(N^2) space, I just haven't gotten to the point of being able to pull DP solutions out of the air.