Implement prison-cells-after-n-days
https://leetcode.com/problems/prison-cells-after-n-days/

time: 208ms
time-rank: 6.06%
time-complexity: O(N)

space: 14.8MB
space-rank: 5.89%
space-complexity: O(N)

The DP solution is good in terms of reusing previous calculations,
but this problem can actually be solved in O(1) time by calculating
the period for each possible input value, and then simply calculating
the remainder of the cycle. It's obvious that there is a cycle, but
it's not obvious that it would be beneficial to use it without
simply a straight DP answer.