Implement linked-list-cycle-ii
https://leetcode.com/problems/linked-list-cycle-ii

time: 4ms
time-rank: 98.90%
time-complexity: O(N)

space: 7.8MB
space-rank: 89.71%
space-complexity: O(1)

This solution adds no extra space but destroys the input.
While the cost of checking if a node is visited is still O(1), this solution
is faster than an unordered_set<LinkedList*> lookup because no hashing is
performed, values are likely still in cache, and it is simply a const
comparison.

Fixes #228

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>