refactor util/TreeNode
This change uses the results of serialize-and-deserialize-binary-tree-test.cpp
to serialize and deserialize trees in O(N) time and O(N) space rather than
O(N^2) time and O(N^2) space.

TreeNode_cleanup() is a no-op at the moment because it causes a number of
double free errors. I think I solved this previously be caching all of
the allocated TreeNode instances in a set and only calling delete when the
instance was still in the set.

Fixes #275

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>