Implement count-unique-characters-of-all-substrings-of-a-given-string
name: count-unique-characters-of-all-substrings-of-a-given-string
url: https://leetcode.com/problems/count-unique-characters-of-all-substrings-of-a-given-string
difficulty: hard

time:
time-rank:
time-complexity: O(N^2)

space:
space-rank:
space-complexity: O(N^3)

Fixes #281

This solution was not accepted by LeetCode due to memory usage. At some point, I'll need to come back and
study it, because it can apparently be done in O(N) time and O(1) space.

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>