Implement tstick
name: tstick
url: This is not a leetcode problem
difficulty: 3

This is a testsuite that demonstrates how to convert between RTOS ticks
and struct timespec. It currently works for signed 64-bit ticks where
the value 0 means "no wait", the value -1 means "wait forever", and the
maximum tick value is INT64_MAX.

This implementation does not consider absolute timepoints for the
conversion between struct timespec and ticks, and all time values should
be considered durations only.

Signed-off-by: Chris Friedt <chrisfriedt@gmail.com>